move point light project to lighting effect
