+++ modules/core/src/lib/effect.js
@@ -7,4 +7,6 @@ export default class Effect {
   }
 
   prepare() {}
+
+  getParameters() {}
 }
